requirements.txt
Flask
werkzeug
opencv-python
numpy
tensorflow
Pillow
easyocr
google-generativeai
python-dotenv
gunicorn
protobuf
